Course/world discovery and alt mech for keeping watched rider in sync.
There is a REST API for player states.  I can't use it too aggressively
or I get 429d but it's very useful for finding out which world the user
is in (or not in the game at all).  It's also handy for keeping track
of who the in-game user is watching when they not in the same nearby
vicinity.