relay session mgmt.
reconnect and disconnect working as well as ability to call disconnect,
followed by connect.

The login session expiration is used to automate reconnects but I still
need to work on reconnects when the network is partitioned.  Should be
pretty straight fwd but I would also like to make sure it has some
throttling and backoff if required.