Wrap electron RPC with JSON
Turns out the elctron structured clone strikes again.  It was not
converging Number subclasses to a usable primative value.  Also it was
sending NaN which JSON doesn't support, making web and electron have
different behavior.  Lastly, JSON beats structured clone in performance
everytime, so this should be faster (but is unbenched)