Checkpoint: chasing windmills.
Working towards fixing missing chats but this is really just some
cleanup and extra debug / experiments.

I'm able to repro broken chat by disconnecting and reconnecting a few
times.  If the TCP server used for a new connection is different we drop
chats.  It seems to require a timeout on the server side.  So we most
likely need to indicate to the servers that the old TCP conn is no
longer in use.  First attempts at using the "leave" endpoint are not
promising but I may have already poisoned the well and should continue
trying to frame the repro cases.