Use /relay/session/refresh instead of building a new tcp socket
Fix bug for upstream change that would trigger multipleLogins condition
for our normal session refresh.  There is a very simple session refresh
API that I've not used before. It allows us to use the same TCP sock
for as long as we want and also to reuse the same relay-id for as long
as we want (help fix chat bugs maybe?)

Fixes #106