More mem savings with inlineNP
Add option for `disableInline(NP|XP)Resize`.  If we don't need resize
support then we don't need to save the stream of np and xp terms.

It just so happens we don't need resize support in S4Z so this is
another full set of data points we don't need to collect anymore.