Stats perf enhancements / extended data life when accessed
Improve performance of periodized rolls data collection and formatting.
We are still leaving some good perf on the table because we use numeric
keys in objects which as it turns out murders performance.

Keep track of athlete-data accessed timestamp for garbage collection. So
if you're using analysis on an athlete that left the game you can
basically evaluate them indefinately as long as you have a tab open on
their data. I.e. calling getAthleteData().