Remove support for env.getRoadCurvePath(_, __, reverse=true)
This is ambigious for a road no matter what.  Do not allow.
Restore compat with RoadPath.toReversed -> CurvePath().

Reversing a road MUST return a CurvePath path object as it destroys
meaning for the roadTime indexing scheme.

Luckily this commit is mostly cleanup and prevention because of a bug
in stats related to gap distance calcs we would never actually send
the `reverse` argument as tru to getRoadCurvePath() and also we never
depended on this being set with our argument structure (start is always
less than end, except in cases where distance should be 0).