Add env.getRouteRoadSections
Design still in flux, but this is the meat of the work previously
handled by segment projection.  Final design will probably eliminate or
simplify the segment projection and use this core func as the basis
instead.  Segment projection will take place in the UI code and
then in a different form on the backend for TBD "To-Next" field(s).