Refactor lapish -> dataslice, add event-slices
Track data for events outside of laps.  Trigging a lap for an
event is okay but if you want to do laps inside an event it's
immediately corrupted.  It's also a bit overlaiden trying to
evaluate which laps are event based.  So add a 3rd dataslice (prev
lapish) for tracking events seperately.