Fix surgical template rendering for select nodes.
There are some side effects when replacing select options one by one
because as the fragment is drained of it's children it mutates the
selected state of it's remaining children.  These states end up
overriding the target node if they are next in the queue for
replacement.  I.e. The select value would change erroneously.